Numerous changes to implement sequencing functionality
Sequencing is now working in very basic form. Pygrbl can load a file, and when play is pressed, it will start to send it to grbl, line by line. As grbl returns position and line number, Pygrbl updates its display, following the tool tip in the OpenGL window, and highlighting the line currently being machined in the editor window. If the user presses "Pause" an emergency stop gets sent. 

This is rickety as anything, and needs a lot of work. In particular:
- it should start from any line in the source code, not just at the start
- pause should pause machining, not stop everything dead
- stepping of machining lines should be implemented
- grbl returns status of the machine, as well as x, y, z and line number, and that status should be displayed
- if the user types a line for immediate action, and it includes a line number, that line number will be highlighted in the editor, even though the editor is being ignored
- in general, if grbl falls over, it leads to a lockup in pygrbl. That badly needs work

But having said all that, it sequences correctly, and I'm off to see if it cuts metal.

Signed-off-by: bobgates