Added logging code and simplified queueing
There was a problem with queueing to grbl that was related to backlash compensation (see corresponding code changes to grbl). In the debugging effort, I've cleaned up and simplified the way that pygrbl ticks and accesses grbl, and I've also added logging code to pygrbl and grblserial. 

Code is now simulating my g-code test piece what appears to be correctly.

Still to do:
- Allow sequencing to stop and restart.

Signed-off-by: bobgates