Fixed comment problem and line numbering
Nothing like cutting real metal to show issues. The code was taking out commented lines in the plot display, but still sending them to grbl which was choking. That's fixed. Also, pygrbl replaces a line number, if there is one, with the number of the actual line in the file. Except it was only replacing, not adding, so g-code files without line numbers weren't being accurately displayed. That's also fixed.

**** There is still a problem with stopping a running g-code program. The pause button activates an emergency stop on grbl, which stops okay. But when grbl is restarted it runs a whole bunch of commands already in the queue before accepting new commands, which causes lots of problems.

I'm impatient to see metal cut, so will ignore this problem for now and work on it when I finish the first test.

Signed-off-by: bobgates