Changed g-code editor to QTableWidget
Lots of small changes, but the g-code editor is now a QTableWidget. It already has some status highlighting as the run occurs, some of which works correctly. It'll also get breakpoints and other symbols fairly soon.

The system is currently working well with the latest version of grbl: it send instructions, keeps track of mistakes, and generally behaves itself. Needs quite a lot more work to polish the user friendliness. In the near future: change the queing so that I can make changes to g-code in the editor and see the changes go out to the machine.

Signed-off-by: bobgates